requirements.txt
streamlit==1.25.0
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
spacy==3.6.1
PyPDF2==3.0.1
python-docx==0.8.11
docx2txt==0.8
matplotlib==3.7.2
seaborn==0.12.2
wordcloud==1.9.2